pyproject.toml
[project]
name = "spoken-diary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "deepgram-sdk==3.4.0",
    "gradio>=6.5.1",
    "loguru>=0.7.3",
    "pcloud>=2.0.1",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
